+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v4.5.1
 
 - Add option to change app icon [#828](https://github.com/hushenghao/AndroidEasterEggs/issues/828)
+- Add analog clock widget dial styles
 - Update App icon [#826](https://github.com/hushenghao/AndroidEasterEggs/issues/826)
 - Update splash screen icon
 - Android 16 Landroid add progress notifications as live updates [#829](https://github.com/hushenghao/AndroidEasterEggs/issues/829)